requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg[binary]==3.2.9
matplotlib==3.10.1
